made get_all_files_list exposed publicly, updated tests
