+++ s3parq/__init__.py
@@ -2,6 +2,7 @@ __version__ = "2.1.4a"
 
 from s3parq.fetch_parq import fetch
 from s3parq.fetch_parq import fetch_diff
+from s3parq.fetch_parq import get_all_files_list
 from s3parq.fetch_parq import get_max_partition_value
 from s3parq.publish_parq import publish
 from s3parq.session_helper import SessionHelper
